Three hazards this branch made reachable, or left behind
Adversarial verification (docs/investigations/2026-08-19-verify-kinect-camera-path.md)
found three defects that the createFilter fix and the LP64 pass either created
or failed to finish. All three are codegen-neutral on PPC; the first two are
native-only correctness.

1. EQEffect.cpp's FILTER was 0x1010 bytes; filterdesign.cpp's is 0x1014.

   Every field EQEffect *reads* lines up, so the two declarations look
   compatible -- but copyresults() (filterdesign.cpp:353) ends with
   `out->numpoles = zplane.numpoles`, a 4-byte store at 0x1010. That was inert
   while `createFilter` was an unresolved extern "C" reference. *This branch is
   what makes the write reachable*, so the out-of-bounds store on
   SetParameter's local `FILTER filter;` has to be closed in the same change
   that opens it.

   It is latent rather than live only because EQEffect::SetParameters has 0
   callers and 0 vtable references in dc3-native today.

   Adding the trailing `int numpoles;` is free on PPC, measured not assumed:
   ?SetParameter@EQEffect@@QAAXHM@Z is 573 instructions / 213 mismatches
   (123 diff_arg, 6 diff_op, 25 replace, 32 insert, 27 delete) and frame
   `stwu r1,-0x10f0(r1)` both before and after -- byte-identical objdiff output.
   The 0x10 frame delta against the target is callee-save GPR counts
   (__savegprlr_25 vs __savegprlr_23), not this struct. sizeof now agrees at
   0x1014 on both sides.

2. LiveCameraInput.cpp:267-268 kept unsigned right-hand sides.

   The branch fixed pointer *truncation* in the four UpdateFrom* bodies but not
   pointer *wrap*: `((pitch>>1) - 640) * 2` and `(mPitch>>2) - 320` are unsigned
   int subtractions. The sibling UpdateFromColorBufferClip already casts to int
   before the multiply for destStride; line 267 never did, and neither did
   line 385's `srcPtr += (srcPitch - 320)` -- the same shape in the function
   the verifier cited as getting it right.

   On 32-bit PPC the wrap steps the pointer correctly backwards; on LP64 it
   zero-extends to ~4e9 and jumps ~8 GB / ~16 GB forward.

   PPC-neutral, measured on all four bodies at the verifier's baselines:
   UpdateFromColorBuffer 75/26, UpdateFromColorBufferClip 158/71,
   UpdateFromDepthBuffer 87/22, UpdateFromDepthBufferClip 117/38 -- unchanged.

   Negative control, so the flat numbers are not vacuous: perturbing that same
   line to `- 641` moved normalized 95.5 -> 95.4 and added `off:+1` to
   mismatch [41] `subi`; restoring the file reproduced 95.5 exactly.

3. POSIX FileEnumerate turned a crash into a silent empty result.

   Un-guarding RecursePatternInternal made it depend on File_Native.cpp's
   enumerator, which is missing File_Win.cpp:107-111's `qualified == "."`
   special case. FileQualifiedFilename prepends gNativeDataDir (default "."),
   so `buf` came back data-dir-qualified while the pattern was not, and
   FileMatch (File.cpp:229-251) requires a literal prefix match up to the first
   wildcard. Any pattern not starting with a wildcard therefore enumerated
   *empty rather than failing* -- reaching ShaderProgram.cpp:53 and
   obj/Utl.cpp:482,517.

Ruler for every number above: objdiff instruction-table counts via the MCP
orchestrator (functionRelocDiffs=none), not headline percentages.

+++ src/system/gesture/LiveCameraInput.cpp
@@ -264,8 +264,16 @@ void LiveCameraInput::TextureStore::UpdateFromColorBuffer(LiveCameraInput *cam) 
                 ((unsigned short *)destPtr)[1] = YUVtoRGB(pixel & 0xff, cr, cb);
                 destPtr += 4;
             }
-            destPtr += ((pitch >> 1) - 640) * 2;
-            srcPtr += (lockedRect.mPitch >> 2) - 320;
+            // (int) casts, matching UpdateFromColorBufferClip's destStride
+            // below. Both right-hand sides are `unsigned int` subtractions: if
+            // the texel pitch is narrower than 640 texels (or the source pitch
+            // narrower than 320 dwords) the result wraps to ~4e9 instead of
+            // going negative. On 32-bit PPC that wrap still steps the pointer
+            // correctly backwards, so this is codegen-neutral there; on the
+            // LP64 native build it zero-extends and jumps ~8 GB / ~16 GB
+            // forward.
+            destPtr += (int)((pitch >> 1) - 640) * 2;
+            srcPtr += (int)((lockedRect.mPitch >> 2) - 320);
         }
         D3DCubeTexture_UnlockRect((D3DCubeTexture *)bufferData, (D3DCUBEMAP_FACES)0, 0);
     }
@@ -386,7 +394,7 @@ void LiveCameraInput::TextureStore::UpdateFromColorBufferClip(
                 }
             }
             destPtr += destStride;
-            srcPtr += (srcPitch - 320);
+            srcPtr += (int)(srcPitch - 320);  // same unsigned-wrap hazard as line 267
         }
         D3DCubeTexture_UnlockRect((D3DCubeTexture *)bufferData, (D3DCUBEMAP_FACES)0, 0);
     }
